move styling to where it makes sense
