+++ src/frontend/src/components/FilterPanel/style.ts
@@ -3,10 +3,8 @@ import {
   Chip,
   CommonThemeProps,
   ComplexFilter,
-  fontCapsXxxs,
   getColors,
   getSpaces,
-  InputDropdown,
 } from "czifui";
 
 export interface ExtraProps extends CommonThemeProps {
@@ -28,25 +26,6 @@ export const StyledFilterPanel = styled("div", {
       display: ${isOpen ? "block" : "none"};
       padding: ${spaces?.xl}px;
       width: 240px;
-    `;
-  }}
-`;
-
-export const StyledInputDropdown = styled(InputDropdown)`
-  ${fontCapsXxxs}
-  padding: 0;
-  text-align: left;
-`;
-
-export const StyledFilterWrapper = styled.div`
-  ${(props) => {
-    const spaces = getSpaces(props);
-    return `
-      margin: ${spaces?.l}px 0;
-
-      &:first-child {
-        margin: 0;
-      }
     `;
   }}
 `;
